raylib::core

Module audio

Contains code related to audio. RaylibAudio plays sounds and music.

Structs

AudioStream

Music

RaylibAudio

This token is used to indicate audio is initialized. It’s also used to create Wave, Sound, Music, AudioStream, and SoundAlias. All of those have a lifetime that is bound to RaylibAudio. The compiler will disallow you from using them without ensuring that the RaylibAudio is present while doing so.

RaylibAudioInitError

Sound

SoundAlias

Wave

WaveSamples

Traits

AudioSample

A marker trait specifying an audio sample (u8, i16, or f32).