raylib::core::color

Struct Color

#[repr(C)]
pub struct Color {
    pub r: u8,
    pub g: u8,
    pub b: u8,
    pub a: u8,
}

Fields
r: u8g: u8b: u8a: u8
Implementations

impl Color

pub fn from_hex(color_hex_str: &str) -> Result<Color, ParseIntError>

Get color from HEX RGB string

Arguments

color_hex_str - A string slice, 6 characters long

Example

   use raylib::prelude::*;
    let color_white = Color::from_hex("FFFFFF").unwrap();
    let color_black = Color::from_hex("000000").unwrap();
     
   assert_eq!(color_black, Color::BLACK);
   assert_eq!(color_white, Color::WHITE);

pub const fn new(r: u8, g: u8, b: u8, a: u8) -> Color

pub fn color_to_int(&self) -> i32

Returns hexadecimal value for a Color

pub fn color_normalize(&self) -> Vector4

Returns color normalized as float [0..1]

pub fn color_to_hsv(&self) -> Vector3

Returns HSV values for a Color

pub fn color_from_hsv(hue: f32, saturation: f32, value: f32) -> Color

Returns a Color from HSV values

pub fn color_from_normalized(normalized: Vector4) -> Color

Returns color from normalized values [0..1]

use raylib::prelude::*;
fn main() {    
    assert_eq!(Color::color_from_normalized(Vector4::new(1.0, 1.0, 1.0, 1.0)), Color::new(255, 255, 255, 255));
}

pub fn get_color(hex_value: u32) -> Color

Returns a Color struct from hexadecimal value

pub fn tint(&self, color: Self) -> Self

Get color multiplied with another color

pub fn brightness(&self, factor: f32) -> Self

Get color with brightness correction, brightness factor goes from -1.0f to 1.0f

pub fn contrast(&self, factor: f32) -> Self

Get color with contrast correction, contrast values between -1.0f and 1.0f

pub fn alpha(&self, alpha: f32) -> Self

Get color with alpha applied, alpha goes from 0.0f to 1.0f

pub fn fade(&self, alpha: f32) -> Self

👎Deprecated: Use Color::alpha instead

Get color with alpha applied, alpha goes from 0.0f to 1.0f

pub fn color_alpha_blend(dst: &Color, src: &Color, tint: &Color) -> Color

Color fade-in or fade-out, alpha goes from 0.0f to 1.0f

pub fn is_equal(&self, rhs: impl Into<Color>) -> bool

Check if color is equal to another.

pub fn lerp(&self, rhs: Color, factor: f32) -> Color

Get color lerp interpolation between two colors, factor [0.0f..1.0f]

impl Color

Color constants

pub const INDIANRED: Color

pub const LIGHTCORAL: Color

pub const SALMON: Color

pub const DARKSALMON: Color

pub const LIGHTSALMON: Color

pub const CRIMSON: Color

pub const RED: Color

pub const FIREBRICK: Color

pub const DARKRED: Color

pub const PINK: Color

pub const LIGHTPINK: Color

pub const HOTPINK: Color

pub const DEEPPINK: Color

pub const MEDIUMVIOLETRED: Color

pub const PALEVIOLETRED: Color

pub const CORAL: Color

pub const TOMATO: Color

pub const ORANGERED: Color

pub const DARKORANGE: Color

pub const ORANGE: Color

pub const GOLD: Color

pub const YELLOW: Color

pub const LIGHTYELLOW: Color

pub const LEMONCHIFFON: Color

pub const LIGHTGOLDENRODYELLOW: Color

pub const PAPAYAWHIP: Color

pub const MOCCASIN: Color

pub const PEACHPUFF: Color

pub const PALEGOLDENROD: Color

pub const KHAKI: Color

pub const DARKKHAKI: Color

pub const LAVENDER: Color

pub const THISTLE: Color

pub const PLUM: Color

pub const VIOLET: Color

pub const ORCHID: Color

pub const FUCHSIA: Color

pub const MAGENTA: Color

pub const MEDIUMORCHID: Color

pub const MEDIUMPURPLE: Color

pub const REBECCAPURPLE: Color

pub const BLUEVIOLET: Color

pub const DARKVIOLET: Color

pub const DARKORCHID: Color

pub const DARKMAGENTA: Color

pub const PURPLE: Color

pub const DARKPURPLE: Color

pub const INDIGO: Color

pub const SLATEBLUE: Color

pub const DARKSLATEBLUE: Color

pub const MEDIUMSLATEBLUE: Color

pub const GREENYELLOW: Color

pub const CHARTREUSE: Color

pub const LAWNGREEN: Color

pub const LIME: Color

pub const LIMEGREEN: Color

pub const PALEGREEN: Color

pub const LIGHTGREEN: Color

pub const MEDIUMSPRINGGREEN: Color

pub const SPRINGGREEN: Color

pub const MEDIUMSEAGREEN: Color

pub const SEAGREEN: Color

pub const FORESTGREEN: Color

pub const GREEN: Color

pub const DARKGREEN: Color

pub const YELLOWGREEN: Color

pub const OLIVEDRAB: Color

pub const OLIVE: Color

pub const DARKOLIVEGREEN: Color

pub const MEDIUMAQUAMARINE: Color

pub const DARKSEAGREEN: Color

pub const LIGHTSEAGREEN: Color

pub const DARKCYAN: Color

pub const TEAL: Color

pub const AQUA: Color

pub const CYAN: Color

pub const LIGHTCYAN: Color

pub const PALETURQUOISE: Color

pub const AQUAMARINE: Color

pub const TURQUOISE: Color

pub const MEDIUMTURQUOISE: Color

pub const DARKTURQUOISE: Color

pub const CADETBLUE: Color

pub const STEELBLUE: Color

pub const LIGHTSTEELBLUE: Color

pub const POWDERBLUE: Color

pub const LIGHTBLUE: Color

pub const SKYBLUE: Color

pub const LIGHTSKYBLUE: Color

pub const DEEPSKYBLUE: Color

pub const DODGERBLUE: Color

pub const CORNFLOWERBLUE: Color

pub const ROYALBLUE: Color

pub const BLUE: Color

pub const MEDIUMBLUE: Color

pub const DARKBLUE: Color

pub const NAVY: Color

pub const MIDNIGHTBLUE: Color

pub const CORNSILK: Color

pub const BLANCHEDALMOND: Color

pub const BISQUE: Color

pub const NAVAJOWHITE: Color

pub const WHEAT: Color

pub const BURLYWOOD: Color

pub const TAN: Color

pub const ROSYBROWN: Color

pub const SANDYBROWN: Color

pub const GOLDENROD: Color

pub const DARKGOLDENROD: Color

pub const PERU: Color

pub const CHOCOLATE: Color

pub const SADDLEBROWN: Color

pub const SIENNA: Color

pub const BROWN: Color

pub const DARKBROWN: Color

pub const MAROON: Color

pub const WHITE: Color

pub const SNOW: Color

pub const HONEYDEW: Color

pub const MINTCREAM: Color

pub const AZURE: Color

pub const ALICEBLUE: Color

pub const GHOSTWHITE: Color

pub const WHITESMOKE: Color

pub const SEASHELL: Color

pub const BEIGE: Color

pub const OLDLACE: Color

pub const FLORALWHITE: Color

pub const IVORY: Color

pub const ANTIQUEWHITE: Color

pub const LINEN: Color

pub const LAVENDERBLUSH: Color

pub const MISTYROSE: Color

pub const GAINSBORO: Color

pub const LIGHTGRAY: Color

pub const SILVER: Color

pub const DARKGRAY: Color

pub const GRAY: Color

pub const DIMGRAY: Color

pub const LIGHTSLATEGRAY: Color

pub const SLATEGRAY: Color

pub const DARKSLATEGRAY: Color

pub const BLACK: Color

pub const BLANK: Color

pub const RAYWHITE: Color