raylib::core::drawing

Trait RaylibShaderModeExt

pub trait RaylibShaderModeExt
where
    Self: Sized,
{
    // Provided methods
    fn begin_shader_mode<'a>(
        &'a mut self,
        shader: &'a mut Shader,
    ) -> RaylibShaderMode<'_, Self> { ... }
    fn draw_shader_mode<'a>(
        &'a mut self,
        shader: &'a mut Shader,
        func: impl FnMut(RaylibShaderMode<'_, Self>),
    ) { ... }
}

Provided Methods

fn begin_shader_mode<'a>( &'a mut self, shader: &'a mut Shader, ) -> RaylibShaderMode<'_, Self>

Begin custom shader drawing. Prefer using the closure version, RaylibShaderModeExt::draw_shader_mode. This version returns a handle that calls raylib_sys::EndShaderMode at the end of the scope and is provided as a fallback incase you run into issues with closures(such as lifetime or performance reasons)

fn draw_shader_mode<'a>( &'a mut self, shader: &'a mut Shader, func: impl FnMut(RaylibShaderMode<'_, Self>), )

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.