raylib::core::drawing

Trait RaylibTextureModeExt

pub trait RaylibTextureModeExt
where
    Self: Sized,
{
    // Provided methods
    fn begin_texture_mode<'a>(
        &'a mut self,
        _: &RaylibThread,
        framebuffer: &'a mut RenderTexture2D,
    ) -> RaylibTextureMode<'_, Self> { ... }
    fn draw_texture_mode<'a>(
        &'a mut self,
        _: &RaylibThread,
        framebuffer: &'a mut RenderTexture2D,
        func: impl FnMut(RaylibTextureMode<'_, Self>),
    ) { ... }
}

Provided Methods

fn begin_texture_mode<'a>( &'a mut self, _: &RaylibThread, framebuffer: &'a mut RenderTexture2D, ) -> RaylibTextureMode<'_, Self>

Begin drawing to render texture. Prefer using the closure version, RaylibTextureModeExt::draw_texture_mode . This version returns a handle that calls raylib_sys::EndTextureMode at the end of the scope and is provided as a fallback incase you run into issues with closures(such as lifetime or performance reasons)

fn draw_texture_mode<'a>( &'a mut self, _: &RaylibThread, framebuffer: &'a mut RenderTexture2D, func: impl FnMut(RaylibTextureMode<'_, Self>), )

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.