raylib::core::drawing

Trait RaylibVRModeExt

pub trait RaylibVRModeExt
where
    Self: Sized,
{
    // Provided methods
    fn begin_vr_stereo_mode<'a>(
        &'a mut self,
        _: &RaylibThread,
        vr_config: &'a mut VrStereoConfig,
    ) -> RaylibVRMode<'_, Self> { ... }
    fn draw_vr_stereo_mode<'a>(
        &'a mut self,
        vr_config: &'a mut VrStereoConfig,
        func: impl FnMut(RaylibVRMode<'_, Self>),
    ) { ... }
}

Provided Methods

fn begin_vr_stereo_mode<'a>( &'a mut self, _: &RaylibThread, vr_config: &'a mut VrStereoConfig, ) -> RaylibVRMode<'_, Self>

Begin stereo rendering (requires VR simulator). Prefer using the closure version, RaylibVRModeExt::draw_vr_stereo_mode . This version returns a handle that calls raylib_sys::EndVrStereoMode at the end of the scope and is provided as a fallback incase you run into issues with closures(such as lifetime or performance reasons)

fn draw_vr_stereo_mode<'a>( &'a mut self, vr_config: &'a mut VrStereoConfig, func: impl FnMut(RaylibVRMode<'_, Self>), )

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.