raylib::core::math

Struct Vector3

#[repr(C)]
pub struct Vector3 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Fields
x: f32y: f32z: f32
Implementations

impl Vector3

pub const fn new(x: f32, y: f32, z: f32) -> Vector3

Returns a new Vector3 with specified components.

pub fn up() -> Vector3

pub fn forward() -> Vector3

pub fn right() -> Vector3

pub fn left() -> Vector3

pub fn zero() -> Vector3

Returns a new Vector3 with all components set to zero.

pub fn one() -> Vector3

Returns a new Vector3 with all components set to one.

pub fn cross(&self, v: Vector3) -> Vector3

Returns a new Vector3 containing the cross product between self and vector v.

pub fn perpendicular(&self) -> Vector3

Returns a new Vector3 perpendicular to self.

pub fn length(&self) -> f32

Calculates the vector length.

pub fn dot(&self, v: Vector3) -> f32

Calculates the dot product with vector v.

pub fn distance_to(&self, v: Vector3) -> f32

Calculates the distance towards vector v.

pub fn scale(&mut self, scale: f32)

Scales the vector by multiplying both components by scale.

pub fn scale_by(&self, scale: f32) -> Vector3

Returns a new Vector3 with components scaled by scale.

pub fn normalize(&mut self)

Normalizes the current vector.

pub fn normalized(&self) -> Vector3

Returns a new Vector3 with normalized components from the current vector.

pub fn ortho_normalize(&mut self, v: &mut Vector3)

Normalizes and changes both self and v to be orthogonal to eachother.

pub fn transform(&mut self, mat: Matrix)

Transforms the current vector using Matrix mat.

pub fn transform_with(&self, mat: Matrix) -> Vector3

Returns a new Vector3 containing components transformed by Matrix mat.

pub fn rotate(&mut self, q: Quaternion)

Rotates the current vector using Quaternion q.

pub fn rotate_by(&self, q: Quaternion) -> Vector3

Returns a new Vector3 with components rotated by Quaternion q.

pub fn lerp(&self, v: Vector3, amount: f32) -> Vector3

Returns a new Vector3 with componenets linearly interpolated by amount towards vector v.

pub fn reflect(&mut self, normal: Vector3)

Reflects the current vector from normal.

pub fn reflect_from(&self, normal: Vector3) -> Vector3

Returns a new Vector3 reflected from the current vector using normal.

pub fn min(&self, v: Vector3) -> Vector3

Returns a new Vector3 containing the minimum of each corresponding component.

pub fn max(&self, v: Vector3) -> Vector3

Returns a new Vector3 containing the maximum of each corresponding component.

pub fn barycenter(&self, a: Vector3, b: Vector3, c: Vector3) -> Vector3

Returns barycenter coordinates (u, v, w) from point p (current vector) with respect to triangle (a, b, c).

pub fn to_array(&self) -> [f32; 3]

Returns a 3-length f32 array containing components [x, y, z] of the current vector.

pub fn clamp(&self, num: Range<f32>) -> Vector3

Returns a new Vector3 with componenets clamp to a certain interval.