raylib::core::math

Type Alias Quaternion

pub type Quaternion = Vector4;

Aliased Type

#[repr(C)]
pub struct Quaternion {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub w: f32,
}

Fields
x: f32y: f32z: f32w: f32
Implementations

impl Quaternion

pub const fn new(x: f32, y: f32, z: f32, w: f32) -> Quaternion

Returns a new Quaternion with specified components.

pub fn identity() -> Quaternion

Returns the identity quaternion.

pub fn from_vec3_pair(from: Vector3, to: Vector3) -> Quaternion

Returns quaternion based on the rotation from one vector to another.

pub fn from_matrix(mat: Matrix) -> Quaternion

Returns a quaternion for a given rotation matrix.

pub fn to_matrix(&self) -> Matrix

Returns a rotation matrix for the current quaternion.

pub fn from_euler(pitch: f32, yaw: f32, roll: f32) -> Quaternion

Returns a quaternion equivalent to Euler angles.

pub fn to_euler(&self) -> Vector3

Returns a vector containing Euler angles in radians (roll, pitch, yaw), based on the current quaternion.

pub fn from_axis_angle(axis: Vector3, angle: f32) -> Quaternion

Returns rotation quaternion for an axis and angle (in radians).

pub fn to_axis_angle(&self) -> (Vector3, f32)

Returns a 2-tuple containing the axis (Vector3) and angle (f32 in radians) for the current quaternion.

pub fn length(&self) -> f32

Computes the length of the current quaternion.

pub fn normalized(&self) -> Quaternion

Returns a normalized version of the current quaternion.

pub fn inverted(&self) -> Quaternion

Returns an inverted version of the current quaternion.

pub fn lerp(&self, q: Quaternion, amount: f32) -> Quaternion

Calculates linear interpolation between current and q quaternions.

pub fn nlerp(&self, q: Quaternion, amount: f32) -> Quaternion

Calculates slerp-optimized interpolation between current and q quaternions.

pub fn slerp(&self, q: Quaternion, amount: f32) -> Quaternion

Calculates spherical linear interpolation between current and q quaternions.

pub fn transform(&self, mat: Matrix) -> Quaternion

Returns a transformed version of the current quaternion given a transformation matrix.

pub fn clamp(&self, num: Range<f32>) -> Quaternion

Returns a new Quaternion with componenets clamp to a certain interval.