Struct raylib::core::models::WeakMesh

#[repr(transparent)]
pub struct WeakMesh(_);

Implementations

impl WeakMesh

pub unsafe fn unwrap(self) -> Mesh

Take the raw ffi type. Must manually free memory by calling the proper unload function

impl WeakMesh

pub fn to_raw(self) -> Mesh

returns the unwrapped raylib-sys object

pub unsafe fn from_raw(raw: Mesh) -> Self

converts raylib-sys object to a “safe” version. Make sure to call this function from the thread the resource was created.