raylib::core

Struct RaylibHandle

pub struct RaylibHandle(/* private fields */);

The main interface into the Raylib API.

This is the way in which you will use the vast majority of Raylib’s functionality. A RaylibHandle can be constructed using the init_window function or through a RaylibBuilder obtained with the init function.

Implementations

impl RaylibHandle

pub fn load_automation_event_list( &self, file_name: Option<PathBuf>, ) -> AutomationEventList

pub fn set_automation_event_list(&self, l: &mut AutomationEventList)

pub fn set_automation_event_base_frame(&self, b: i32)

pub fn start_automation_event_recording(&self)

pub fn stop_automation_event_recording(&self)

impl RaylibHandle

pub fn set_trace_log_callback( &mut self, cb: fn(TraceLogLevel, &str), ) -> Result<(), SetLogError<'_>>

👎Deprecated: Decoupled from RaylibHandle. Use set_trace_log_callback instead.

Set custom trace log

pub fn set_save_file_data_callback( &mut self, cb: fn(&str, &[u8]) -> bool, ) -> Result<(), SetLogError<'_>>

👎Deprecated: Decoupled from RaylibHandle. Use set_save_file_data_callback instead.

Set custom file binary data saver

pub fn set_load_file_data_callback<'b>( &mut self, cb: fn(&str) -> Vec<u8>, ) -> Result<(), SetLogError<'_>>

👎Deprecated: Decoupled from RaylibHandle. Use set_load_file_data_callback instead.

Set custom file binary data loader

Whatever you return from your callback will be intentionally leaked as Raylib is relied on to free it.

pub fn set_save_file_text_callback( &mut self, cb: fn(&str, &str) -> bool, ) -> Result<(), SetLogError<'_>>

👎Deprecated: Decoupled from RaylibHandle. Use set_save_file_text_callback instead.

Set custom file text data saver

pub fn set_load_file_text_callback( &mut self, cb: fn(&str) -> String, ) -> Result<(), SetLogError<'_>>

👎Deprecated: Decoupled from RaylibHandle. Use set_load_file_text_callback instead.

Set custom file text data loader

Whatever you return from your callback will be intentionally leaked as Raylib is relied on to free it.

pub fn set_audio_stream_callback( &mut self, stream: AudioStream<'_>, cb: fn(&[u8]), ) -> Result<(), SetLogError<'_>>

👎Deprecated: Decoupled from RaylibHandle. Use set_audio_stream_callback instead.

Audio thread callback to request new data

impl RaylibHandle

pub fn update_camera(&self, camera: &mut Camera3D, mode: CameraMode)

Updates camera position for selected mode.

pub fn update_camera_pro( &self, camera: &mut Camera3D, movement: Vector3, rotation: Vector3, zoom: f32, )

impl RaylibHandle

Seems like all draw commands must be issued from the main thread

pub fn begin_drawing(&mut self, _: &RaylibThread) -> RaylibDrawHandle<'_>

Setup canvas (framebuffer) to start drawing. Prefer using the closure version, RaylibHandle::draw. This version returns a handle that calls raylib_sys::EndDrawing at the end of the scope and is provided as a fallback incase you run into issues with closures(such as lifetime or performance reasons)

pub fn draw(&mut self, _: &RaylibThread, func: impl FnMut(RaylibDrawHandle<'_>))

impl RaylibHandle

pub fn is_file_dropped(&self) -> bool

Checks if a file has been dropped into the window.

pub fn is_file_extension<A>(&self, file_name: A, file_ext: A) -> bool
where A: Into<OsString>,

Checks a file’s extension.

pub fn application_directory(&self) -> String

Get the directory of the running application.

pub fn get_file_length<A>(&self, filename: A) -> i32
where A: Into<OsString>,

Get file length in bytes.

Errors

This function will return an error if the supplied bytes contain an internal 0 byte. The NulError returned will contain the bytes as well as the position of the nul byte.

pub fn is_path_file<A>(&self, filename: A) -> bool
where A: Into<OsString>,

Check if a given path is a file or a directory

Errors

This function will return an error if the supplied bytes contain an internal 0 byte. The NulError returned will contain the bytes as well as the position of the nul byte.

pub fn load_directory_files<A>(&self, dir_path: A) -> FilePathList
where A: Into<OsString>,

Load directory filepaths

pub fn load_directory_files_ex<A>( &self, dir_path: A, filter: String, scan_sub_dirs: bool, ) -> FilePathList
where A: Into<OsString>,

Load directory filepaths with extension filtering and recursive directory scan

pub fn load_dropped_files(&self) -> DroppedFilePathList

Check if a file has been dropped into window

impl RaylibHandle

pub fn is_key_pressed(&self, key: KeyboardKey) -> bool

Detect if a key has been pressed once.

pub fn is_key_pressed_repeat(&self, key: KeyboardKey) -> bool

Check if a key has been pressed again

pub fn is_key_down(&self, key: KeyboardKey) -> bool

Detect if a key is being pressed.

pub fn is_key_released(&self, key: KeyboardKey) -> bool

Detect if a key has been released once.

pub fn is_key_up(&self, key: KeyboardKey) -> bool

Detect if a key is NOT being pressed.

pub fn get_key_pressed(&mut self) -> Option<KeyboardKey>

Gets latest key pressed.

pub fn get_key_pressed_number(&mut self) -> Option<u32>

Gets latest key pressed.

pub fn get_char_pressed(&mut self) -> Option<char>

Gets latest char (unicode) pressed

pub fn set_exit_key(&mut self, key: Option<KeyboardKey>)

Sets a custom key to exit program (default is ESC).

pub fn is_gamepad_available(&self, gamepad: i32) -> bool

Detect if a gamepad is available.

pub fn get_gamepad_name(&self, gamepad: i32) -> Option<String>

Returns gamepad internal name id.

pub fn is_gamepad_button_pressed( &self, gamepad: i32, button: GamepadButton, ) -> bool

Detect if a gamepad button has been pressed once.

pub fn is_gamepad_button_down( &self, gamepad: i32, button: GamepadButton, ) -> bool

Detect if a gamepad button is being pressed.

pub fn is_gamepad_button_released( &self, gamepad: i32, button: GamepadButton, ) -> bool

Detect if a gamepad button has been released once.

pub fn is_gamepad_button_up(&self, gamepad: i32, button: GamepadButton) -> bool

Detect if a gamepad button is NOT being pressed.

pub fn get_gamepad_button_pressed(&self) -> Option<GamepadButton>

Gets the last gamepad button pressed.

pub fn get_gamepad_axis_count(&self, gamepad: i32) -> i32

Returns gamepad axis count for a gamepad.

pub fn get_gamepad_axis_movement(&self, gamepad: i32, axis: GamepadAxis) -> f32

Returns axis movement value for a gamepad axis.

pub fn is_mouse_button_pressed(&self, button: MouseButton) -> bool

Detect if a mouse button has been pressed once.

pub fn is_mouse_button_down(&self, button: MouseButton) -> bool

Detect if a mouse button is being pressed.

pub fn is_mouse_button_released(&self, button: MouseButton) -> bool

Detect if a mouse button has been released once.

pub fn is_mouse_button_up(&self, button: MouseButton) -> bool

Detect if a mouse button is NOT being pressed.

pub fn get_mouse_x(&self) -> i32

Returns mouse position X.

pub fn get_mouse_y(&self) -> i32

Returns mouse position Y.

pub fn get_mouse_position(&self) -> Vector2

Returns mouse position.

pub fn get_mouse_delta(&self) -> Vector2

Returns mouse delta between frames.

pub fn set_mouse_position(&mut self, position: impl Into<Vector2>)

Sets mouse position.

pub fn set_mouse_offset(&mut self, offset: impl Into<Vector2>)

Sets mouse offset.

pub fn set_mouse_scale(&mut self, scale_x: f32, scale_y: f32)

Sets mouse scaling.

pub fn get_mouse_wheel_move(&self) -> f32

Get mouse wheel movement for X or Y, whichever is larger

pub fn get_mouse_wheel_move_v(&self) -> Vector2

Get mouse wheel movement for both X and Y

pub fn get_touch_x(&self) -> i32

Returns touch position X for touch point 0 (relative to screen size).

pub fn get_touch_y(&self) -> i32

Returns touch position Y for touch point 0 (relative to screen size).

pub fn get_touch_position(&self, index: u32) -> Vector2

Returns touch position XY for a touch point index (relative to screen size).

pub fn set_gestures_enabled(&self, gesture_flags: u32)

Enables a set of gestures using flags.

pub fn set_gamepad_mappings(&self, bind: &[c_char]) -> i32

Set internal gamepad mappings (SDL_GameControllerDB)

pub fn set_gamepad_vibration( &mut self, gamepad: i32, left_motor: f32, right_motor: f32, duration: f32, )

Set gamepad vibration for both motors

pub fn is_gesture_detected(&self, gesture: Gesture) -> bool

Checks if a gesture have been detected.

pub fn get_gesture_detected(&self) -> Gesture

Gets latest detected gesture.

pub fn get_touch_point_id(&self, index: u32) -> i32

Get touch point identifier for given index

pub fn get_touch_point_count(&self) -> u32

Gets touch points count.

pub fn get_gesture_hold_duration(&self) -> f32

Gets gesture hold time in milliseconds.

pub fn get_gesture_drag_vector(&self) -> Vector2

Gets gesture drag vector.

pub fn get_gesture_drag_angle(&self) -> f32

Gets gesture drag angle.

pub fn get_gesture_pinch_vector(&self) -> Vector2

Gets gesture pinch delta.

pub fn get_gesture_pinch_angle(&self) -> f32

Gets gesture pinch angle.

impl RaylibHandle

pub fn set_trace_log(&self, types: TraceLogLevel)

Set the current threshold (minimum) log level

pub fn trace_log(&self, msg_type: TraceLogLevel, text: &str)

Writes a trace log message (Log::INFO, Log::WARNING, Log::ERROR, Log::DEBUG).

impl RaylibHandle

pub fn load_random_sequence<'a>( &self, num: Range<i32>, count: u32, ) -> RandomSequence<'a>

Load random values sequence, no values repeated

pub fn load_image_from_screen(&self, _: &RaylibThread) -> Image

Load pixels from the screen into a CPU image

pub fn take_screenshot(&mut self, _: &RaylibThread, filename: &str)

Takes a screenshot of current screen (saved a .png)

pub fn get_random_value<T: From<i32>>(&self, num: Range<i32>) -> T

Returns a random value between min and max (both included)

use raylib::*;
fn main() {
    let (mut rl, thread) = ...;
    let r = rl.get_random_value(0, 10);
    println!("random value: {}", r);
}

pub fn set_random_seed(&mut self, seed: u32)

Set the seed for random number generation

impl RaylibHandle

pub fn load_model( &mut self, _: &RaylibThread, filename: &str, ) -> Result<Model, Error>

Loads model from files (mesh and material).

pub fn load_model_from_mesh( &mut self, _: &RaylibThread, mesh: WeakMesh, ) -> Result<Model, Error>

pub fn load_model_animations( &mut self, _: &RaylibThread, filename: &str, ) -> Result<Vec<ModelAnimation>, Error>

pub fn update_model_animation( &mut self, _: &RaylibThread, model: impl AsMut<Model>, anim: impl AsRef<ModelAnimation>, frame: i32, )

pub fn update_model_animation_bones( &mut self, _: &RaylibThread, model: impl AsMut<Model>, anim: impl AsRef<ModelAnimation>, frame: i32, )

impl RaylibHandle

pub fn load_material_default(&self, _: &RaylibThread) -> WeakMaterial

pub unsafe fn unload_material( &mut self, _: &RaylibThread, material: WeakMaterial, )

Weak materials will leak memeory if they are not unlaoded Unload material from GPU memory (VRAM)

pub unsafe fn unload_model(&mut self, _: &RaylibThread, model: WeakModel)

Weak models will leak memeory if they are not unlaoded Unload model from GPU memory (VRAM)

pub unsafe fn unload_model_animation( &mut self, _: &RaylibThread, model_animation: WeakModelAnimation, )

Weak model_animations will leak memeory if they are not unlaoded Unload model_animation from GPU memory (VRAM)

pub unsafe fn unload_mesh(&mut self, _: &RaylibThread, mesh: WeakMesh)

Weak meshs will leak memeory if they are not unlaoded Unload mesh from GPU memory (VRAM)

impl RaylibHandle

pub fn load_shader( &mut self, _: &RaylibThread, vs_filename: Option<&str>, fs_filename: Option<&str>, ) -> Shader

Loads a custom shader and binds default locations.

pub fn load_shader_from_memory( &mut self, _: &RaylibThread, vs_code: Option<&str>, fs_code: Option<&str>, ) -> Shader

Loads shader from code strings and binds default locations.

impl RaylibHandle

pub fn get_codepoint_count(text: &str) -> i32

pub fn unload_font(&mut self, font: WeakFont)

pub fn load_font( &mut self, _: &RaylibThread, filename: &str, ) -> Result<Font, Error>

Loads font from file into GPU memory (VRAM).

pub fn load_font_ex( &mut self, _: &RaylibThread, filename: &str, font_size: i32, chars: Option<&str>, ) -> Result<Font, Error>

Loads font from file with extended parameters. Supplying None for chars loads the entire character set.

pub fn load_font_from_image( &mut self, _: &RaylibThread, image: &Image, key: impl Into<Color>, first_char: i32, ) -> Result<Font, Error>

Load font from Image (XNA style)

pub fn load_font_from_memory( &mut self, _: &RaylibThread, file_type: &str, file_data: &[u8], font_size: i32, chars: Option<&str>, ) -> Result<Font, Error>

Load font data from a given memory buffer. file_type refers to the extension, e.g. “.ttf”. You can pass Some(…) to chars to get the desired charaters, or None to get the whole set.

pub fn load_font_data( &mut self, data: &[u8], font_size: i32, chars: Option<&str>, sdf: i32, ) -> Option<RSliceGlyphInfo>

Loads font data for further use (see also Font::from_data). Now supports .tiff

impl RaylibHandle

pub fn get_font_default(&self) -> WeakFont

Gets the default font.

pub fn measure_text(&self, text: &str, font_size: i32) -> i32

Measures string width in pixels for default font.

pub fn set_text_line_spacing(&self, spacing: i32)

impl RaylibHandle

pub fn load_texture( &mut self, _: &RaylibThread, filename: &str, ) -> Result<Texture2D, Error>

Loads texture from file into GPU memory (VRAM).

pub fn load_texture_cubemap( &mut self, _: &RaylibThread, image: &Image, layout: CubemapLayout, ) -> Result<Texture2D, Error>

Load cubemap from image, multiple image cubemap layouts supported

pub fn load_texture_from_image( &mut self, _: &RaylibThread, image: &Image, ) -> Result<Texture2D, Error>

Loads texture from image data.

pub fn load_render_texture( &mut self, _: &RaylibThread, width: u32, height: u32, ) -> Result<RenderTexture2D, Error>

Loads texture for rendering (framebuffer).

impl RaylibHandle

pub unsafe fn unload_texture( &mut self, _: &RaylibThread, texture: WeakTexture2D, )

Weak Textures will leak memeory if they are not unlaoded Unload textures from GPU memory (VRAM)

pub unsafe fn unload_render_texture( &mut self, _: &RaylibThread, texture: WeakRenderTexture2D, )

Weak RenderTextures will leak memeory if they are not unlaoded Unload RenderTextures from GPU memory (VRAM)

impl RaylibHandle

pub fn load_vr_stereo_config( &mut self, _: &RaylibThread, device: impl Into<VrDeviceInfo>, ) -> VrStereoConfig

impl RaylibHandle

pub fn get_clipboard_text(&self) -> Result<String, Utf8Error>

Get clipboard text content

pub fn set_clipboard_text(&mut self, text: &str) -> Result<(), NulError>

Set clipboard text content

impl RaylibHandle

pub fn get_screen_to_world_ray( &self, mouse_position: impl Into<Vector2>, camera: impl Into<Camera>, ) -> Ray

Get a ray trace from screen position (i.e mouse)

pub fn get_screen_to_world_ray_ex( &self, mouse_position: impl Into<Vector2>, camera: impl Into<Camera>, width: i32, height: i32, ) -> Ray

Get a ray trace from screen position (i.e mouse) in a viewport

pub fn get_world_to_screen( &self, position: impl Into<Vector3>, camera: impl Into<Camera>, ) -> Vector2

Returns the screen space position for a 3d world space position

pub fn get_world_to_screen2D( &self, position: impl Into<Vector2>, camera: impl Into<Camera2D>, ) -> Vector2

Returns the screen space position for a 2d camera world space position

pub fn get_world_to_screen_ex( &self, position: impl Into<Vector3>, camera: impl Into<Camera>, width: i32, height: i32, ) -> Vector2

Returns size position for a 3d world space position

pub fn get_screen_to_world2D( &self, position: impl Into<Vector2>, camera: impl Into<Camera2D>, ) -> Vector2

Returns the world space position for a 2d camera screen space position

impl RaylibHandle

pub fn set_target_fps(&mut self, fps: u32)

Set target FPS (maximum)

pub fn get_fps(&self) -> u32

Returns current FPS

pub fn get_frame_time(&self) -> f32

Returns time in seconds for last frame drawn

pub fn get_time(&self) -> f64

Returns elapsed time in seconds since InitWindow()

impl RaylibHandle

pub fn window_should_close(&self) -> bool

Checks if KEY_ESCAPE or Close icon was pressed. Do not call on web unless you are compiling with asyncify.

pub fn is_window_ready(&self) -> bool

Checks if window has been initialized successfully.

pub fn maximize_window(&mut self)

Set window state: maximized, if resizable

pub fn minimize_window(&mut self)

Set window state: minimized, if resizable

pub fn restore_window(&mut self)

Set window state: not minimized/maximized

pub fn is_window_maximized(&self) -> bool

Check if window is currently maximized

pub fn is_window_minimized(&self) -> bool

Checks if window has been minimized (or lost focus).

pub fn is_window_resized(&self) -> bool

Checks if window has been resized.

pub fn is_window_hidden(&self) -> bool

Checks if window has been hidden.

pub fn is_window_fullscreen(&self) -> bool

Returns whether or not window is in fullscreen mode

pub fn is_window_focused(&self) -> bool

pub fn get_window_scale_dpi(&self) -> Vector2

Check if window is currently focused (only PLATFORM_DESKTOP)

pub fn is_cursor_on_screen(&self) -> bool

Check if cursor is on the current screen.

pub fn set_mouse_cursor(&self, cursor: MouseCursor)

Set mouse cursor

pub fn toggle_fullscreen(&mut self)

Toggles fullscreen mode (only on desktop platforms).

pub fn set_window_state(&mut self, state: WindowState)

Set window configuration state using flags

pub fn clear_window_state(&mut self, state: WindowState)

Clear window configuration state flags

pub fn get_window_state(&self) -> WindowState

Get the window config state

pub fn set_window_icon(&mut self, image: impl AsRef<Image>)

Sets icon for window (only on desktop platforms).

pub fn set_window_icons(&mut self, images: &mut [Image])

pub fn set_window_title(&self, _: &RaylibThread, title: &str)

Sets title for window (only on desktop platforms).

pub fn set_window_position(&mut self, x: i32, y: i32)

Sets window position on screen (only on desktop platforms).

pub fn set_window_monitor(&mut self, monitor: i32)

Sets monitor for the current window (fullscreen mode).

pub fn set_window_min_size(&mut self, width: i32, height: i32)

Sets minimum window dimensions (for FLAG_WINDOW_RESIZABLE).

pub fn set_window_max_size(&mut self, width: i32, height: i32)

Sets maximum window dimensions (for FLAG_WINDOW_RESIZABLE).

pub fn set_window_size(&mut self, width: i32, height: i32)

Sets window dimensions.

pub fn set_window_opacity(&mut self, opacity: f32)

Set window opacity, value opacity is between 0.0 and 1.0

pub fn get_render_width(&self) -> i32

Get current render width which is equal to screen width * dpi scale

pub fn get_render_height(&self) -> i32

Get current render width which is equal to screen height * dpi scale

pub fn get_screen_width(&self) -> i32

Get current screen height which is equal to screen height * dpi scale

pub fn get_screen_height(&self) -> i32

Gets current screen height.

pub fn get_window_position(&self) -> Vector2

Get window position

pub fn toggle_borderless_windowed(&self)

pub fn set_window_focused(&self)

impl RaylibHandle

pub fn show_cursor(&mut self)

Shows mouse cursor.

pub fn hide_cursor(&mut self)

Hides mouse cursor.

pub fn is_cursor_hidden(&self) -> bool

Checks if mouse cursor is not visible.

pub fn enable_cursor(&mut self)

Enables mouse cursor (unlock cursor).

pub fn disable_cursor(&mut self)

Disables mouse cursor (lock cursor).

pub unsafe fn get_window_handle(&mut self) -> *mut c_void

Get native window handle

impl RaylibHandle

Global gui modification functions

pub fn gui_enable(&mut self)

Enable gui controls (global state)

pub fn gui_disable(&mut self)

Disable gui controls (global state)

pub fn gui_lock(&mut self)

Lock gui controls (global state)

pub fn gui_unlock(&mut self)

Unlock gui controls (global state)

pub fn gui_fade(&mut self, color: Color, alpha: f32) -> Color

pub fn gui_set_state(&mut self, state: GuiState)

Set gui state (global state)

pub fn gui_get_state(&mut self) -> GuiState

Get gui state (global state)

pub fn gui_set_font(&mut self, font: impl AsRef<Font>)

Set gui custom font (global state)

pub fn gui_get_font(&mut self) -> WeakFont

Get gui custom font (global state)

pub fn gui_set_style( &mut self, control: GuiControl, property: impl GuiProperty, value: i32, )

Set one style property

pub fn gui_get_style( &mut self, control: GuiControl, property: impl GuiProperty, ) -> i32

Get one style property

pub fn gui_load_style(&mut self, filename: &str)

Load style file (.rgs)

pub fn gui_load_style_default(&mut self)

Load style default over global style

pub fn gui_enable_tooltip(&mut self)

Enable gui tooltips (global state)

pub fn gui_disable_tooltip(&mut self)

Disable gui tooltips (global state)

pub fn gui_set_tooltip(&mut self, tooltip: &str)

Set tooltip string