Module raylib::core::text

Text and Font related functions Text manipulation functions are super unsafe so use rust String functions

Structs

CharInfo

Font

RSliceCharInfo

WeakFont

Enums

FontLoadEx

Parameters for Font::load_font_ex

Traits

RaylibFont

Functions

gen_image_font_atlas

Generates image font atlas using chars info. Sets a pointer to an array of rectangles raylib allocated that MUST manually be freed. Good luck freeing it safely though ;)

get_glyph_index

Gets index position for a unicode character on font.

measure_text

Measures string width in pixels for default font.

measure_text_ex

Measures string width in pixels for font.