raylib::core

Module window

Window manipulation functions

Structs

MonitorInfo

WindowState

Functions

get_camera_matrix

Returns camera transform matrix (view matrix)

get_camera_matrix2D

Returns camera 2D transform matrix (view matrix)

get_current_monitor

get_current_monitor_index

get_monitor_count

Get number of connected monitors

get_monitor_height

Get height of monitor Only checks that monitor index is in range in debug mode

get_monitor_info

Gets the attributes of the monitor as well as the name fails if monitor name is not a utf8 string

get_monitor_name

Get name of monitor Only checks that monitor index is in range in debug mode

get_monitor_physical_height

Get physical height of monitor Only checks that monitor index is in range in debug mode

get_monitor_physical_width

Get physical width of monitor Only checks that monitor index is in range in debug mode

get_monitor_position

Get position of monitor Only checks that monitor index is in range in debug mode

get_monitor_refresh_rate

Get specified monitor refresh rate

get_monitor_width

Get width of monitor Only checks that monitor index is in range in debug mode