raylib

Module ease

Easing and interpolation helpers.

Tween provides a simple way in which to interpolate a single f32 value. The various easing functions contained within this module may be used with it.

Structs

Tween

A manager for a tween on a single f32 value.

Functions

back_in

back_in_out

back_out

bounce_in

bounce_in_out

bounce_out

circ_in

circ_in_out

circ_out

cubic_in

cubic_in_out

cubic_out

elastic_in

elastic_in_out

elastic_out

expo_in

expo_in_out

expo_out

linear_in

linear_in_out

linear_none

linear_out

quad_in

quad_in_out

quad_out

sine_in

sine_in_out

sine_out

Type Aliases

EaseFn

The type alias used for all easing functions.