Struct raylib::core::RaylibHandle

pub struct RaylibHandle(_);

The main interface into the Raylib API.

This is the way in which you will use the vast majority of Raylib’s functionality. A RaylibHandle can be constructed using the init_window function or through a RaylibBuilder obtained with the init function.

Implementations

impl RaylibHandle

pub fn set_camera_mode(&mut self, camera: impl Into<Camera3D>, mode: CameraMode)

Sets camera mode.

pub fn update_camera(&self, camera: &mut Camera3D)

Updates camera position for selected mode.

pub fn set_camera_pan_control(&mut self, pan_key: KeyboardKey)

Sets camera pan key to combine with mouse movement (free camera).

pub fn set_camera_alt_control(&mut self, alt_key: KeyboardKey)

Sets camera alt key to combine with mouse movement (free camera).

pub fn set_camera_smooth_zoom_control(&mut self, sz_key: KeyboardKey)

Sets camera smooth zoom key to combine with mouse (free camera).

pub fn set_camera_move_controls(
&mut self,
front_key: KeyboardKey,
back_key: KeyboardKey,
right_key: KeyboardKey,
left_key: KeyboardKey,
up_key: KeyboardKey,
down_key: KeyboardKey
)

Sets camera move controls (1st person and 3rd person cameras).

impl RaylibHandle

Seems like all draw commands must be issued from the main thread

pub fn begin_drawing(&mut self, _: &RaylibThread) -> RaylibDrawHandle<'_>

Setup canvas (framebuffer) to start drawing

impl RaylibHandle

pub fn is_file_dropped(&self) -> bool

Checks if a file has been dropped into the window.

pub fn get_dropped_files(&self) -> Vec<String>

Gets dropped filenames.

pub fn clear_dropped_files(&mut self)

Clears dropped files paths buffer.

impl RaylibHandle

pub fn is_key_pressed(&self, key: KeyboardKey) -> bool

Detect if a key has been pressed once.

pub fn is_key_down(&self, key: KeyboardKey) -> bool

Detect if a key is being pressed.

pub fn is_key_released(&self, key: KeyboardKey) -> bool

Detect if a key has been released once.

pub fn is_key_up(&self, key: KeyboardKey) -> bool

Detect if a key is NOT being pressed.

pub fn get_key_pressed(&mut self) -> Option<KeyboardKey>

Gets latest key pressed.

pub fn get_key_pressed_number(&mut self) -> Option<u32>

Gets latest key pressed.

pub fn set_exit_key(&mut self, key: Option<KeyboardKey>)

Sets a custom key to exit program (default is ESC).

pub fn is_gamepad_available(&self, gamepad: i32) -> bool

Detect if a gamepad is available.

pub fn is_gamepad_name(&self, gamepad: i32, name: &str) -> bool

Checks gamepad name (if available).

pub fn get_gamepad_name(&self, gamepad: i32) -> Option<String>

Returns gamepad internal name id.

pub fn is_gamepad_button_pressed(
&self,
gamepad: i32,
button: GamepadButton
) -> bool

Detect if a gamepad button has been pressed once.

pub fn is_gamepad_button_down(
&self,
gamepad: i32,
button: GamepadButton
) -> bool

Detect if a gamepad button is being pressed.

pub fn is_gamepad_button_released(
&self,
gamepad: i32,
button: GamepadButton
) -> bool

Detect if a gamepad button has been released once.

pub fn is_gamepad_button_up(&self, gamepad: i32, button: GamepadButton) -> bool

Detect if a gamepad button is NOT being pressed.

pub fn get_gamepad_button_pressed(&self) -> Option<GamepadButton>

Gets the last gamepad button pressed.

pub fn get_gamepad_axis_count(&self, gamepad: i32) -> i32

Returns gamepad axis count for a gamepad.

pub fn get_gamepad_axis_movement(&self, gamepad: i32, axis: GamepadAxis) -> f32

Returns axis movement value for a gamepad axis.

pub fn is_mouse_button_pressed(&self, button: MouseButton) -> bool

Detect if a mouse button has been pressed once.

pub fn is_mouse_button_down(&self, button: MouseButton) -> bool

Detect if a mouse button is being pressed.

pub fn is_mouse_button_released(&self, button: MouseButton) -> bool

Detect if a mouse button has been released once.

pub fn is_mouse_button_up(&self, button: MouseButton) -> bool

Detect if a mouse button is NOT being pressed.

pub fn get_mouse_x(&self) -> i32

Returns mouse position X.

pub fn get_mouse_y(&self) -> i32

Returns mouse position Y.

pub fn get_mouse_position(&self) -> Vector2

Returns mouse position.

pub fn set_mouse_position(&mut self, position: impl Into<Vector2>)

Sets mouse position.

pub fn set_mouse_offset(&mut self, offset: impl Into<Vector2>)

Sets mouse offset.

pub fn set_mouse_scale(&mut self, scale_x: f32, scale_y: f32)

Sets mouse scaling.

pub fn get_mouse_wheel_move(&self) -> f32

Returns mouse wheel movement Y.

pub fn get_touch_x(&self) -> i32

Returns touch position X for touch point 0 (relative to screen size).

pub fn get_touch_y(&self) -> i32

Returns touch position Y for touch point 0 (relative to screen size).

pub fn get_touch_position(&self, index: u32) -> Vector2

Returns touch position XY for a touch point index (relative to screen size).

pub fn set_gestures_enabled(&self, gesture_flags: u32)

Enables a set of gestures using flags.

pub fn is_gesture_detected(&self, gesture: Gestures) -> bool

Checks if a gesture have been detected.

pub fn get_gesture_detected(&self) -> Gestures

Gets latest detected gesture.

pub fn get_touch_points_count(&self) -> u32

Gets touch points count.

pub fn get_gesture_hold_duration(&self) -> f32

Gets gesture hold time in milliseconds.

pub fn get_gesture_drag_vector(&self) -> Vector2

Gets gesture drag vector.

pub fn get_gesture_drag_angle(&self) -> f32

Gets gesture drag angle.

pub fn get_gesture_pinch_vector(&self) -> Vector2

Gets gesture pinch delta.

pub fn get_gesture_pinch_angle(&self) -> f32

Gets gesture pinch angle.

impl RaylibHandle

pub fn get_screen_data(&mut self, _: &RaylibThread) -> Image

pub fn take_screenshot(&mut self, _: &RaylibThread, filename: &str)

Takes a screenshot of current screen (saved a .png)

impl RaylibHandle

pub fn load_model(
&mut self,
_: &RaylibThread,
filename: &str
) -> Result<Model, String>

Loads model from files (mesh and material).

pub fn load_model_from_mesh(
&mut self,
_: &RaylibThread,
mesh: WeakMesh
) -> Result<Model, String>

pub fn load_model_animations(
&mut self,
_: &RaylibThread,
filename: &str
) -> Result<Vec<ModelAnimation>, String>

pub fn update_model_animation(
&mut self,
_: &RaylibThread,
model: impl AsMut<Model>,
anim: impl AsRef<ModelAnimation>,
frame: i32
)

impl RaylibHandle

pub fn load_material_default(&self, _: &RaylibThread) -> WeakMaterial

pub unsafe fn unload_material(
&mut self,
_: &RaylibThread,
material: WeakMaterial
)

Weak materials will leak memeory if they are not unlaoded Unload material from GPU memory (VRAM)

pub unsafe fn unload_model(&mut self, _: &RaylibThread, model: WeakModel)

Weak models will leak memeory if they are not unlaoded Unload model from GPU memory (VRAM)

pub unsafe fn unload_model_animation(
&mut self,
_: &RaylibThread,
model_animation: WeakModelAnimation
)

Weak model_animations will leak memeory if they are not unlaoded Unload model_animation from GPU memory (VRAM)

pub unsafe fn unload_mesh(&mut self, _: &RaylibThread, mesh: WeakMesh)

Weak meshs will leak memeory if they are not unlaoded Unload mesh from GPU memory (VRAM)

impl RaylibHandle

pub fn load_shader(
&mut self,
_: &RaylibThread,
vs_filename: Option<&str>,
fs_filename: Option<&str>
) -> Result<Shader, String>

Loads a custom shader and binds default locations.

pub fn load_shader_from_memory(
&mut self,
_: &RaylibThread,
vs_code: Option<&str>,
fs_code: Option<&str>
) -> Shader

Loads shader from code strings and binds default locations.

impl RaylibHandle

pub fn unload_font(&mut self, font: WeakFont)

pub fn load_font(
&mut self,
_: &RaylibThread,
filename: &str
) -> Result<Font, String>

Loads font from file into GPU memory (VRAM).

pub fn load_font_ex(
&mut self,
_: &RaylibThread,
filename: &str,
font_size: i32,
chars: FontLoadEx<'_>
) -> Result<Font, String>

Loads font from file with extended parameters.

pub fn load_font_from_image(
&mut self,
_: &RaylibThread,
image: &Image,
key: impl Into<Color>,
first_char: i32
) -> Result<Font, String>

Load font from Image (XNA style)

pub fn load_font_data(
&mut self,
data: &[u8],
font_size: i32,
chars: Option<&[i32]>,
sdf: i32
) -> Option<RSliceCharInfo>

Loads font data for further use (see also Font::from_data). Now supports .tiff

impl RaylibHandle

pub fn get_font_default(&self) -> WeakFont

Gets the default font.

impl RaylibHandle

pub fn load_texture(
&mut self,
_: &RaylibThread,
filename: &str
) -> Result<Texture2D, String>

Loads texture from file into GPU memory (VRAM).

pub fn load_texture_cubemap(
&mut self,
_: &RaylibThread,
image: &Image,
layout: CubemapLayout
) -> Result<Texture2D, String>

Load cubemap from image, multiple image cubemap layouts supported

pub fn load_texture_from_image(
&mut self,
_: &RaylibThread,
image: &Image
) -> Result<Texture2D, String>

Loads texture from image data.

pub fn load_render_texture(
&mut self,
_: &RaylibThread,
width: u32,
height: u32
) -> Result<RenderTexture2D, String>

Loads texture for rendering (framebuffer).

impl RaylibHandle

pub unsafe fn unload_texture(
&mut self,
_: &RaylibThread,
texture: WeakTexture2D
)

Weak Textures will leak memeory if they are not unlaoded Unload textures from GPU memory (VRAM)

pub unsafe fn unload_render_texture(
&mut self,
_: &RaylibThread,
texture: WeakRenderTexture2D
)

Weak RenderTextures will leak memeory if they are not unlaoded Unload RenderTextures from GPU memory (VRAM)

impl RaylibHandle

pub fn load_vr_stereo_config(
&mut self,
_: &RaylibThread,
device: impl Into<VrDeviceInfo>
) -> VrStereoConfig

impl RaylibHandle

pub fn get_clipboard_text(&self) -> Result<String, Utf8Error>

Get clipboard text content

pub fn set_clipboard_text(&mut self, text: &str) -> Result<(), NulError>

Set clipboard text content

impl RaylibHandle

pub fn get_mouse_ray(
&self,
mouse_position: impl Into<Vector2>,
camera: impl Into<Camera>
) -> Ray

Returns a ray trace from mouse position

pub fn get_world_to_screen(
&self,
position: impl Into<Vector3>,
camera: impl Into<Camera>
) -> Vector2

Returns the screen space position for a 3d world space position

pub fn get_world_to_screen2D(
&self,
position: impl Into<Vector2>,
camera: impl Into<Camera2D>
) -> Vector2

Returns the screen space position for a 2d camera world space position

pub fn get_world_to_screen_ex(
&self,
position: impl Into<Vector3>,
camera: impl Into<Camera>,
width: i32,
height: i32
) -> Vector2

Returns size position for a 3d world space position

pub fn get_screen_to_world2D(
&self,
position: impl Into<Vector2>,
camera: impl Into<Camera2D>
) -> Vector2

Returns the world space position for a 2d camera screen space position

impl RaylibHandle

pub fn set_target_fps(&mut self, fps: u32)

Set target FPS (maximum)

pub fn get_fps(&self) -> u32

Returns current FPS

pub fn get_frame_time(&self) -> f32

Returns time in seconds for last frame drawn

pub fn get_time(&self) -> f64

Returns elapsed time in seconds since InitWindow()

impl RaylibHandle

pub fn window_should_close(&self) -> bool

Checks if KEY_ESCAPE or Close icon was pressed.

pub fn is_window_ready(&self) -> bool

Checks if window has been initialized successfully.

pub fn is_window_minimized(&self) -> bool

Checks if window has been minimized (or lost focus).

pub fn is_window_resized(&self) -> bool

Checks if window has been resized.

pub fn is_window_hidden(&self) -> bool

Checks if window has been hidden.

pub fn is_window_fullscreen(&self) -> bool

Returns whether or not window is in fullscreen mode

pub fn is_window_focused(&self) -> bool

pub fn get_window_scale_dpi(&self) -> Vector2

Check if window is currently focused (only PLATFORM_DESKTOP)

pub fn is_cursor_on_screen(&self) -> bool

Check if cursor is on the current screen.

pub fn set_mouse_cursor(&self, cursor: MouseCursor)

Set mouse cursor

pub fn toggle_fullscreen(&mut self)

Toggles fullscreen mode (only on desktop platforms).

pub fn set_window_state(&mut self, state: WindowState)

Set window configuration state using flags

pub fn clear_window_state(&mut self, state: WindowState)

Clear window configuration state flags

pub fn get_window_state(&self) -> WindowState

Get the window config state

pub fn set_window_icon(&mut self, image: impl AsRef<Image>)

Sets icon for window (only on desktop platforms).

pub fn set_window_title(&self, _: &RaylibThread, title: &str)

Sets title for window (only on desktop platforms).

pub fn set_window_position(&mut self, x: i32, y: i32)

Sets window position on screen (only on desktop platforms).

pub fn set_window_monitor(&mut self, monitor: i32)

Sets monitor for the current window (fullscreen mode).

pub fn set_window_min_size(&mut self, width: i32, height: i32)

Sets minimum window dimensions (for FLAG_WINDOW_RESIZABLE).

pub fn set_window_size(&mut self, width: i32, height: i32)

Sets window dimensions.

pub fn get_screen_width(&self) -> i32

Gets current screen width.

pub fn get_screen_height(&self) -> i32

Gets current screen height.

pub fn get_window_position(&self) -> Vector2

Get window position

impl RaylibHandle

pub fn show_cursor(&mut self)

Shows mouse cursor.

pub fn hide_cursor(&mut self)

Hides mouse cursor.

pub fn is_cursor_hidden(&self) -> bool

Checks if mouse cursor is not visible.

pub fn enable_cursor(&mut self)

Enables mouse cursor (unlock cursor).

pub fn disable_cursor(&mut self)

Disables mouse cursor (lock cursor).

pub unsafe fn get_window_handle(&mut self) -> *mut c_void

Get native window handle

impl RaylibHandle

Global gui modification functions

pub fn gui_enable(&mut self)

Enable gui controls (global state)

pub fn gui_disable(&mut self)

Disable gui controls (global state)

pub fn gui_lock(&mut self)

Lock gui controls (global state)

pub fn gui_unlock(&mut self)

Unlock gui controls (global state)

pub fn gui_fade(&mut self, alpha: f32)

pub fn gui_set_state(&mut self, state: GuiControlState)

Set gui state (global state)

pub fn gui_get_state(&mut self) -> GuiControlState

Get gui state (global state)

pub fn gui_set_font(&mut self, font: impl AsRef<Font>)

Set gui custom font (global state)

pub fn gui_get_font(&mut self) -> WeakFont

Get gui custom font (global state)

pub fn gui_set_style(&mut self, control: GuiControl, property: i32, value: i32)

Set one style property SHOULD use one of the Gui*Property enums

pub fn gui_get_style(&mut self, control: GuiControl, property: i32) -> i32

Get one style property SHOULD use one of the Gui*Property enums

pub fn gui_load_style(&mut self, filename: Option<&CStr>)

Load style file (.rgs)

pub fn gui_load_style_default(&mut self)

Load style default over global style