Trait raylib::rgui::RaylibDrawGui

pub trait RaylibDrawGui {

    fn gui_enable(&mut self) { ... }

    fn gui_disable(&mut self) { ... }

    fn gui_lock(&mut self) { ... }

    fn gui_unlock(&mut self) { ... }

    fn gui_fade(&mut self, alpha: f32) { ... }

    fn gui_set_state(&mut self, state: GuiControlState) { ... }

    fn gui_get_state(&mut self) -> GuiControlState { ... }

    fn gui_set_font(&mut self, font: impl AsRef<Font>) { ... }

    fn gui_get_font(&mut self) -> WeakFont { ... }

    fn gui_set_style(&mut self, control: GuiControl, property: i32, value: i32) { ... }

    fn gui_get_style(&self, control: GuiControl, property: i32) -> i32 { ... }

    fn gui_load_style(&mut self, filename: Option<&CStr>) { ... }

    fn gui_load_style_default(&mut self) { ... }

    fn gui_window_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        title: Option<&CStr>
    ) -> bool { ... }

    fn gui_group_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>
    ) { ... }

    fn gui_line(&mut self, bounds: impl Into<Rectangle>, text: Option<&CStr>) { ... }

    fn gui_panel(&mut self, bounds: impl Into<Rectangle>) { ... }

    fn gui_scroll_panel(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        content: impl Into<Rectangle>, 
        scroll: impl Into<Vector2>
    ) -> (Rectangle, Vector2) { ... }

    fn gui_label(&mut self, bounds: impl Into<Rectangle>, text: Option<&CStr>) { ... }

    fn gui_button(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>
    ) -> bool { ... }

    fn gui_label_button(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>
    ) -> bool { ... }

    fn gui_image_button(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        texture: impl AsRef<Texture>
    ) -> bool { ... }

    fn gui_image_button_ex(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        texture: impl AsRef<Texture>, 
        tex_source: impl Into<Rectangle>
    ) -> bool { ... }

    fn gui_toggle(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        active: bool
    ) -> bool { ... }

    fn gui_toggle_group(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        active: i32
    ) -> i32 { ... }

    fn gui_check_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        checked: bool
    ) -> bool { ... }

    fn gui_combo_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        active: i32
    ) -> i32 { ... }

    fn gui_dropdown_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        active: &mut i32, 
        edit_mode: bool
    ) -> bool { ... }

    fn gui_spinner(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        value: &mut i32, 
        min_value: i32, 
        max_value: i32, 
        edit_mode: bool
    ) -> bool { ... }

    fn gui_value_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        value: &mut i32, 
        min_value: i32, 
        max_value: i32, 
        edit_mode: bool
    ) -> bool { ... }

    fn gui_text_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        buffer: &mut [u8], 
        edit_mode: bool
    ) -> bool { ... }

    fn gui_text_box_multi(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        buffer: &mut [u8], 
        edit_mode: bool
    ) -> bool { ... }

    fn gui_slider(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text_left: Option<&CStr>, 
        text_right: Option<&CStr>, 
        value: f32, 
        min_value: f32, 
        max_value: f32
    ) -> f32 { ... }

    fn gui_slider_bar(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text_left: Option<&CStr>, 
        text_right: Option<&CStr>, 
        value: f32, 
        min_value: f32, 
        max_value: f32
    ) -> f32 { ... }

    fn gui_progress_bar(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text_left: Option<&CStr>, 
        text_right: Option<&CStr>, 
        value: f32, 
        min_value: f32, 
        max_value: f32
    ) -> f32 { ... }

    fn gui_status_bar(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>
    ) { ... }

    fn gui_dummy_rec(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>
    ) { ... }

    fn gui_scroll_bar(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        value: i32, 
        min_value: i32, 
        max_value: i32
    ) -> i32 { ... }

    fn gui_grid(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        spacing: f32, 
        subdivs: i32
    ) -> Vector2 { ... }

    fn gui_list_view(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        scroll_index: &mut i32, 
        active: i32
    ) -> i32 { ... }

    fn gui_list_view_ex(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: &[&CStr], 
        focus: &mut i32, 
        scroll_index: &mut i32, 
        active: i32
    ) -> i32 { ... }

    fn gui_message_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        text: Option<&CStr>, 
        message: Option<&CStr>, 
        buttons: Option<&CStr>
    ) -> i32 { ... }

    fn gui_text_input_box(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        title: Option<&CStr>, 
        message: Option<&CStr>, 
        buttons: Option<&CStr>, 
        text: &mut Vec<u8>
    ) -> i32 { ... }

    fn gui_color_picker(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        color: impl Into<Color>
    ) -> Color { ... }

    fn gui_icon_text(
        &mut self, 
        icon_id: guiIconName, 
        text: Option<&CStr>
    ) -> String { ... }

    fn gui_color_bar_alpha(
        &mut self, 
        bounds: impl Into<Rectangle>, 
        alpha: f32
    ) -> f32 { ... }

}

Provided methods

fn gui_enable(&mut self)

Enable gui controls (global state)

fn gui_disable(&mut self)

Disable gui controls (global state)

fn gui_lock(&mut self)

Lock gui controls (global state)

fn gui_unlock(&mut self)

Unlock gui controls (global state)

fn gui_fade(&mut self, alpha: f32)

fn gui_set_state(&mut self, state: GuiControlState)

Set gui state (global state)

fn gui_get_state(&mut self) -> GuiControlState

Get gui state (global state)

fn gui_set_font(&mut self, font: impl AsRef<Font>)

Set gui custom font (global state)

fn gui_get_font(&mut self) -> WeakFont

Get gui custom font (global state)

fn gui_set_style(&mut self, control: GuiControl, property: i32, value: i32)

Set one style property SHOULD use one of the Gui*Property enums

fn gui_get_style(&self, control: GuiControl, property: i32) -> i32

Get one style property SHOULD use one of the Gui*Property enums

fn gui_load_style(&mut self, filename: Option<&CStr>)

Load style file (.rgs)

fn gui_load_style_default(&mut self)

Load style default over global style

fn gui_window_box(
&mut self,
bounds: impl Into<Rectangle>,
title: Option<&CStr>
) -> bool

Window Box control, shows a window that can be closed

fn gui_group_box(&mut self, bounds: impl Into<Rectangle>, text: Option<&CStr>)

Group Box control with text name

fn gui_line(&mut self, bounds: impl Into<Rectangle>, text: Option<&CStr>)

Line separator control, could contain text

fn gui_panel(&mut self, bounds: impl Into<Rectangle>)

Panel control, useful to group controls

fn gui_scroll_panel(
&mut self,
bounds: impl Into<Rectangle>,
content: impl Into<Rectangle>,
scroll: impl Into<Vector2>
) -> (Rectangle, Vector2)

Scroll Panel control

fn gui_label(&mut self, bounds: impl Into<Rectangle>, text: Option<&CStr>)

Label control, shows text

fn gui_button(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>
) -> bool

Button control, returns true when clicked

fn gui_label_button(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>
) -> bool

Label button control, show true when clicked

fn gui_image_button(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
texture: impl AsRef<Texture>
) -> bool

Image button control, returns true when clicked

fn gui_image_button_ex(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
texture: impl AsRef<Texture>,
tex_source: impl Into<Rectangle>
) -> bool

Image button extended control, returns true when clicked

fn gui_toggle(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
active: bool
) -> bool

Toggle Button control, returns true when active

fn gui_toggle_group(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
active: i32
) -> i32

Toggle Group control, returns active toggle index

fn gui_check_box(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
checked: bool
) -> bool

Check Box control, returns true when active

fn gui_combo_box(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
active: i32
) -> i32

Combo Box control, returns selected item index

fn gui_dropdown_box(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
active: &mut i32,
edit_mode: bool
) -> bool

Dropdown Box control, returns selected item

fn gui_spinner(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
value: &mut i32,
min_value: i32,
max_value: i32,
edit_mode: bool
) -> bool

Spinner control, returns selected value

fn gui_value_box(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
value: &mut i32,
min_value: i32,
max_value: i32,
edit_mode: bool
) -> bool

Value Box control, updates input text with numbers

fn gui_text_box(
&mut self,
bounds: impl Into<Rectangle>,
buffer: &mut [u8],
edit_mode: bool
) -> bool

Text Box control, updates input text Use at your own risk!!! The allocated vector MUST have enough space for edits.

fn gui_text_box_multi(
&mut self,
bounds: impl Into<Rectangle>,
buffer: &mut [u8],
edit_mode: bool
) -> bool

Text Box control with multiple lines Use at your own risk!!! The allocated vector MUST have a nul terminator.

fn gui_slider(
&mut self,
bounds: impl Into<Rectangle>,
text_left: Option<&CStr>,
text_right: Option<&CStr>,
value: f32,
min_value: f32,
max_value: f32
) -> f32

Slider control, returns selected value

fn gui_slider_bar(
&mut self,
bounds: impl Into<Rectangle>,
text_left: Option<&CStr>,
text_right: Option<&CStr>,
value: f32,
min_value: f32,
max_value: f32
) -> f32

Slider Bar control, returns selected value

fn gui_progress_bar(
&mut self,
bounds: impl Into<Rectangle>,
text_left: Option<&CStr>,
text_right: Option<&CStr>,
value: f32,
min_value: f32,
max_value: f32
) -> f32

Progress Bar control, shows current progress value

fn gui_status_bar(&mut self, bounds: impl Into<Rectangle>, text: Option<&CStr>)

Status Bar control, shows info text

fn gui_dummy_rec(&mut self, bounds: impl Into<Rectangle>, text: Option<&CStr>)

Dummy control for placeholders

fn gui_scroll_bar(
&mut self,
bounds: impl Into<Rectangle>,
value: i32,
min_value: i32,
max_value: i32
) -> i32

Scroll Bar control

fn gui_grid(
&mut self,
bounds: impl Into<Rectangle>,
spacing: f32,
subdivs: i32
) -> Vector2

Grid control

fn gui_list_view(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
scroll_index: &mut i32,
active: i32
) -> i32

List View control, returns selected list item index

fn gui_list_view_ex(
&mut self,
bounds: impl Into<Rectangle>,
text: &[&CStr],
focus: &mut i32,
scroll_index: &mut i32,
active: i32
) -> i32

List View with extended parameters

fn gui_message_box(
&mut self,
bounds: impl Into<Rectangle>,
text: Option<&CStr>,
message: Option<&CStr>,
buttons: Option<&CStr>
) -> i32

Message Box control, displays a message

fn gui_text_input_box(
&mut self,
bounds: impl Into<Rectangle>,
title: Option<&CStr>,
message: Option<&CStr>,
buttons: Option<&CStr>,
text: &mut Vec<u8>
) -> i32

Text Input Box control, ask for text

fn gui_color_picker(
&mut self,
bounds: impl Into<Rectangle>,
color: impl Into<Color>
) -> Color

Color Picker control

fn gui_icon_text(&mut self, icon_id: guiIconName, text: Option<&CStr>) -> String

fn gui_color_bar_alpha(
&mut self,
bounds: impl Into<Rectangle>,
alpha: f32
) -> f32

Color Bar Alpha control NOTE: Returns alpha value normalized [0..1]